refactor: text 클립보드 저장 불완전 로직 개선
